Prepersist client countage action added

+++ src/main/java/pl/WojtekMalka/csvReader/DTO/ClientDTO.java
@@ -8,6 +8,6 @@ public class ClientDTO {
     private String last_name;
     private String birth_date;
     private String phone_no;
-    private Integer client_age;
+    private Long client_age;
 
 }
